+++ src/Libraries/Microsoft.Extensions.AI.Abstractions/ChatCompletion/RequiredChatToolMode.cs
@@ -15,17 +15,17 @@ namespace Microsoft.Extensions.AI;
 public sealed class RequiredChatToolMode : ChatToolMode
 {
     /// <summary>
-    /// Gets the name of a specific <see cref="AIFunction"/> that must be called.
+    /// Gets the name of a specific tool that must be called.
     /// </summary>
     /// <remarks>
-    /// If the value is <see langword="null"/>, any available function can be selected (but at least one must be).
+    /// If the value is <see langword="null"/>, any available tool can be selected (but at least one must be).
     /// </remarks>
     public string? RequiredFunctionName { get; }
 
     /// <summary>
-    /// Initializes a new instance of the <see cref="RequiredChatToolMode"/> class that requires a specific function to be called.
+    /// Initializes a new instance of the <see cref="RequiredChatToolMode"/> class that requires a specific tool to be called.
     /// </summary>
-    /// <param name="requiredFunctionName">The name of the function that must be called.</param>
+    /// <param name="requiredFunctionName">The name of the tool that must be called.</param>
     /// <exception cref="ArgumentException"><paramref name="requiredFunctionName"/> is empty or composed entirely of whitespace.</exception>
     /// <remarks>
     /// <paramref name="requiredFunctionName"/> can be <see langword="null"/>. However, it's preferable to use